Move functions set-up model components to SSeasonTST.utils
